package name refactor and testNG examle
